requirements.txt
streamlit
torch
torchvision
segmentation-models-pytorch
transformers
Pillow
numpy
pandas